!help

+++ lib/discordbot/commands/commands.ex
@@ -1,14 +1,23 @@
 defmodule Discordbot.Commands do
 
   alias DiscordEx.RestClient.Resources.Channel
+  alias Discordbot.Helpers.Message
+
+  def handle(:message_create, data = %{"content" => "!help"}, state = %{rest_client: conn}) do
+    help = Application.get_env(:discordbot, :help)
+    spawn fn -> Channel.delete_message(conn, data["channel_id"], data["id"]) end
+    spawn fn -> Message.dm(conn, data["author"]["id"], help) end
+    {:ok, state}
+  end
 
   def handle(:message_create, data = %{"content" => "!" <> command}, state = %{rest_client: conn}) do
     commands = Application.get_env(:discordbot, :commands)
     command = parse_command(command)
-    if Keyword.has_key?(commands, command.name) do
+    last_command = Map.get(state, :last_command, nil)
+    if Keyword.has_key?(commands, command.name) && command != last_command do
       spawn fn -> Channel.delete_message(conn, data["channel_id"], data["id"]) end
       spawn fn -> Channel.send_message(conn, data["channel_id"], %{content: message(command)}) end
-      {:ok, state}
+      {:ok, Map.update(state, :last_command, command, command)}
     else
       {:no, state}
     end
